isConfidential moved into new class AbstractGroupLayer
